fix possible null 'this' pointer dereference (#1082)
If a nullptr is passed to the second parameter of the Field class's constructor
then there is a nullptr this pointer dereference for the boost::shared_ptr that
is the second parameter as the reshare() method is called.  The second
parameter is a nullptr in BPatch_type::getComponents is the type is and enum
type.  This warning is only diagnosed with -Og (and maybe -O1).