symlite - make symlite/h/SymLite-elf.h publicly consumable
It refers to files in the build tree that aren't available in the
install tree. Although this file is in the install tree, it was never
consumable by users. Updating this is of questionable utility, but it at
least gives us the opportunity to write tests against the public
interface.