dyninstAPI_RT - use new dyninst_library
1. SRC_ASSEMBLY was never used. The assembly files have always been
explicitly listed.

2. We are still linking libdl.so on the static versions of the runtime.
This hasn't been an issue in the past, so we'll leave it.

3. The creation of the lists of source files has redundant checks in it,
but this version is much clearer on what is actually happening.

4. CHECK_C_COMPILER_FLAG_EXTENDED was removed since CMake now supports a
much larger set of compilers and Dyninst has dropped support for HP,
Sun, and XL.

5. The config is much less noisy. There was no reason to determine if
32-bit support would work when the user didn't ask for it. Now, the
config will fail if the user asks for 32-bit libraries and the compiler
can't create them.