Replace DwarfWalker::findString with DwarfWalker::find_call_file
The original function was only ever used to find DW_AT_call_file. During
the transition to libdw, the added check for that attribute made the
rest of the string processing useless. Moreover, the code update also
introduced a bug in checking the return value of dwarf_whatform which
rendered the function completely useless as a generic stringform parser.