Update global ELF ctor/dtor instrumentation for static executables (#1355)
If __libc_csu_init doesn't exist, try to insert the constructor call into the beginning of 'main'. If __libc_csu_fini doesn't exist, try to insert the constructor call into the beginning of 'exit'.

* Use C-style strings for symbol names

There's no need to have a std::string object that incurs the overhead of a constructor/destructor call to just hold a literal that's only visible in the current translation unit.