Require TBB >=2019.9
When building from source, versions before 2019.9 incorrectly set the
version in TBBConfigVersion.cmake. For example, 2018.6 sets the version
to 2018.0 because it uses the TBB_{MAJOR,MINOR}_VERSION from tbb_stddef.h
instead of doing the calculation based on the engineering version.

This also unifies the versions required when compiling with gcc and clang.

TBB 2018.6 was released in Oct 2018 and 2019.9 was released in Oct 2019,
so this just bumps the requirement by just a year even though there are
at least 9 releases in between.