Explicitly delete unusable special member functions
Because there is a user-defined destructor, the compiler will not generate the special member functions (e.g., copy assignment operator). However, we explicitly delete them here to signal that this class is not copyable or movable.  The destructor is also moved to the top of the class to be next to the other special member functions.