Separate out AMDGPU register formatting
Here we extract the logic of displaying multiple registers as
a single operand when doing instruction formatting into a separate helper
function that is only called when the architecture is passed to
Instruction.format or Operand.format;

This change enables the ability get register operand name in two ways:
Consider s[0:3]
1. Without passing in architecture, each register will be printed separately.
That is s0,s1,s2,s3
2. By passing in architecture, the first register will return a string
for the entire group, and the rest will return empty string.
That is s[0:3],"","",""
These empty strings will later be discarded by the
ArchitectureSpecificFormatter functions.