Fix responsibility inversion in Module::finalizeRanges (#1498)
A Module shouldn't modify the Symtab object to which it belongs. It
knows what ranges belong to it (see
Object::fix_global_symbol_modules_static_dwarf), so it can munge them
into a collection of `ModRange`s.

The `ranges_finalized` member was never needed as the only place
it was set was in `finalizeRanges` which is the only place where
`ranges` was emptied.