Revert "Unify naming type units"
It's not clear from the DWARF std at this time if dwarf_formstring
will actually return the signature/MD5 hash for a type unit. Whatever
the old code has done is likely broken, so we don't want to propagate
that here. Likely, the DW_FORM_ref_sig8 parsing isn't being done
correctly overall, and will need to be fixed in the future.

This reverts commit 8a32113cddb2beb7cd5e3fc42221be098a76a602.