Use full filenames when handling Modules (#1501)
A module's name is either a unique string or the pathname of the file
used to create the corresponding DWARF compilation unit. In the latter
case, the 'name' is only the basename of the path which is never
guaranteed to be unique (e.g., /path1/foo.c /path2/foo.c would have the
same 'name').

* Remove useless file creations

These don't correspond to any Module.