Add logicalNot operation to Rose for correctness
Previously rose use invertOp for logical Not.
As the Dyninst interpretation of Rose didn't limit
the size of the input to single bit, inverting a non-zero
value can still result in an non-zero value, causing
incorrect result to be populated.

This PR adds a separate logicalNotOp that
returns 0 given non-zero value
returns 1 given 0