DW_CATCH debugging.
Firewalls the exception throw in the dwarf parser so that
a backtrace is possible.

This catches the 'bad_cast' exception throws.   It seems
they all come from parseMember().

In addition, I noticed that we are not parsing (default error)
the '0' entries on el8 (they don't exist on ubuntu 20/22) and
we might not be parsing dwarf correctly.  It appears to be legal
dwarf that we are doing a default action on, instead of an explicit
action?

There are actually two firewalls, in case an exception is thrown
past the case statement of parse_int() -- which is now _parse_int()
with the exception firewall in parse_int()

The log messages all are tagged DW_CATCH for easy grepping.