Construct a Module from the CU's offset not its PC (#1626)
* Construct a Module from the CU's offset not its PC

The PC value can be non-unique across CUs. For example, they can all
be 0x0 for a PIE binary. The offset of the CU is unique as it's the
location inside of the .debug_info table.

* Use correct lookup when creating module during Object parse

When creating a module during fix_global_symbol_modules_static_dwarf,
the default module covers all ranges, so we need to look for the
exact offset to prevent using the default module every time.

* Use correct offset in DwarfWalker::parseModule

* Lookup DIE location with dwarf_offdie when parsing ranges