Add CMake option to control symbol visibility
Currently, all symbols are hidden by default and
then explicitly made visible. This is the behavior
we want for publishing Dyninst libraries, but unit
testing the internal functionality requires access
to these symbols. This option allows that.

It is explicitly not described in DyninstOptions.cmake
because it is not intended to be used by regular users.