MachRegister: Add interface for collecting and retrieving all registers (#1703)
This makes writing ABI rules much simpler. It could also be useful for
some power users. The registers returned are in the order they are
declared in common/h/registers/<arch>_regs.h.

Names are only save the register when a name is given because
the other constructor is used to convert registers (e.g., getBaseRegister)
and would create duplicates entries since the name isn't changed.