Remove unused annotations (#1717)
* CommonBlockUpPtr

This was added by d4b6dac19 in 2009, but was never written to. The
check on line 243 would always pass and getCblocks would return an
empty vector.

* FunctionLocalVariablesAnno

Its usage was removed by 5ed068167 in 2013.

* Remove FunctionParametersAnno

Its usage was removed by 5ed068167 in 2013.

* Remove IdToSymMap

Its usage was remove dby c42859c30 in 2016.

* Remove ImageFuncUpPtrAnno

Its usage was removed by 5ed068167 in 2013.

* Remove LocalVarUpPtrAnno

It was added by d2e866379 in 2009, but has always been write-only.

* Remove ModuleLineInfoAnno

Its usage was remove by 186761951 in 2014.

* Remove ModuleTypeInfoAnno

Its usage was removed by 186761951 in 2014.

* Remove UserFuncsAnno

Its usage was removed by 74c0f9613 in 2008.

* Remove UserSymbolsAnno

Its usage was removed by 74c0f9613 in 2008.