x86: fix decoding of pextrb, pextrw and add vpextrw with opcode 0x15
Move pextrb and pextrw instructions to the correct position in sse_ter_table
since they are prefixed with 66 and not F2.
Add support for decoding of EVEX prefixed vpextrw instruction.
Make  operands of vpextrw and vpextrb instructions consistent with Intel
manual.
Add support for decoding a variant of vpextrw instruction with opcode 0x15.

Signed-off-by: koltunov <koltunov@ispras.ru>