warning fix: ambiguous-reversed-operator (#1782)
In C++20 and later the signature of operator== should be
T::operator==(const T&) const otherwise it is ambiguous as
C++ 20 and later also looks for matching function with the
arguments reversed and if only one is const then reversing
results in two distinct function signatures that match equally
well.  Some of these were reported as compiler warnings,
others from manual inspection.