Import and update Aarch64 registers from Capstone (#1795)
* Remove Generic Interrupt Controller registers

Capstone has partial support for the Generic Interrupt Controller
Architecture Specification registers. However, these are for embedded
systems only. To reduce the number of registers in Dyninst, we ignore
them for now.

* Move constant format note

This makes it consistent with the other <arch>_regs.h files

* Fix value of IMPLEMENTATION_DEFINED_SYSREG

It's value can't be bigger than 255 since it has to fit in a single
byte.

* Mark hq* as pseudo-registers

They are only used in instruction decoding and will go away once we
use Capstone.

* Update special purpose registers (SPRs)

Capstone doesn't have a separate representations for the nzcv bits
of the pstate flag.

* Mark fpsr as a system register

It's treated that way in the Architecture Reference Manual and in
Capstone.