Add category helpers to Instruction class (#1813)
getCategory() assumes there is only one category per instruction, but
that restriction will be removed when using the Capstone decoder and the
function deprecated. These helpers will make it easier for users to
check behavior without having to know how it's being tracked in
Instruction.

This makes IA_IAPI::hasCFT() more expensive, but it will be much
cheaper as we move more architectures to Capstone.