Move MachRegister -> ROSE register conversion into dataflowAPI (#1842)
The conversion is only ever used in dataflowAPI, so that's where it
belongs. It also simplifies the conversion by directly creating a
RegisterDescriptor which is the register vocabulary for interacting
with ROSE.

* Add logging for unimplemented architectures

There is no `default` case in the switch, so the compiler will warn
about missing enumerators in which case this code should never be
encountered.