Fix register calculation in parse_func::calcUsedRegs for aarch64
The MachRegister ids are no longer guaranteed to be sequential, so
performing arithmetic on them may not be valid. Because 'registerSlot'
uses aarch64Registers_t to represent registers, MachRegisters need to be
transformed using the conversion maps in RegisterConversion-aarch64.C.
Those maps use the the base registers x<N> for GPRs and q<N> for FPRs,
so the registers are correspondingly transformed before being stored in
parse_func::usedRegisters.