[AMDGPU] Change filename of instrumented kernel list
It was a suffix to the output file name that comes from the mutator.
This commit changes it to be a suffix to the input file name. So we can
set up other tools in our scripts to read the instrumented kernel names
without having to worry about the mutator.