Unify entry region construction
This is a slight change for PPC which previously parsed the entire
binary. There's no reason to do that since we know the only region we
need is near the entry point. Additionally, PPC allowed for an entry
region with no calls. That is inconsistent with the libc startup
procedure, so it's removed.