Fix PowerPC branch instruction semantics (#1929)
* Remove blr search in IA_power::isReturn

The instruction decoder now correctly detects subroutine returns based
on the bclr idiom described in the operand decoder there.

* Remove branch hint calculation for BO field

That formula only applies before version 2.0 of the ISA (~2003). We no
longer support anything before Power8 (v2.07), so this is no longer
useful. The hint bits are still present, but there's no need to modify
the mnemonic to reflect that. This will get updated when migrating to
Capstone.

* Ensure branch instructions with BO read/write ctr

Previously, the operand was only added if the instruction wrote to ctr.