Fix detection of conditional jump in slicing (#2062)
* Fix detection of conditional jump in slicing

Currently, in backward slicing, the detection of the conditional jump
assumes both
COND_TAKEN and COND_NOT_TAKEN are present.

However, under current parsing implementation, it could be the case
where only
one of the edges are present when parsing the indirect control flow in
the fallthrough block.

As a result, when backward slicing on the jump table index, the slicer
would miss the fact that
predecessor block ends with a conditional jump instruction, and fail to
compute the bound.

Paritally Fixes #2050