Fix build error in docker/build.sh
I do not know why, but when the install directory ($dest_dir) is
manually created with 'mkdir' instead of letting it be implicitly
created by 'cmake --build', 'ctest' fails to find the tests- even when
I have confirmed they are present.

I've added '--no-tests=error' to ensure that ctest fails when no tests
are run.