cold clone symbols are not fncts, now type CODE
A symbol with a cold clone suffix is not a function entry point, but
is part of the function without the suffix.  The new symbol type
"ST_CODE" represents this type of symbol.

Without this change, cold function symbols are incorrectly treated as
real functions.

ST_CODE symbol types are mapped back to STT_FUNC in ELF

updated description and documentation for clones and version suffixes.